+++ tangelo/linq/circuit.py
@@ -96,6 +96,16 @@ class Circuit:
         """
         return not (self == other)
 
+    def __iter__(self):
+        """Define the iterator. This is useful when iterating through all the
+        gates in a Circuit.
+        """
+        return iter(self._gates)
+
+    def __next__(self):
+        """Define the next function when calling next(Circuit). """
+        return next(self._gates)
+
     @property
     def size(self):
         """The size is the number of gates in the circuit. It is different from
@@ -127,6 +137,10 @@ class Circuit:
         if any MEASURE gate was explicitly added by the user.
         """
         return "MEASURE" in self.counts
+
+    def copy(self):
+        """Return a deepcopy of circuit"""
+        return Circuit(copy.deepcopy(self._gates), n_qubits=self._qubits_simulated, name=self.name)
 
     def add_gate(self, g):
         """Add a new gate to a circuit object and update other fields of the