+++ CHANGELOG.md
@@ -3,6 +3,27 @@
 This file documents the main changes between versions of the code.
 
 
+## [0.3.2] - 2022-08-06
+
+### Added
+
+- Linear Combination of Unitaries (LCU)
+- QEMIST Cloud MI-FNO innregration: interface adjustments
+- iQCC ansatz for VQE
+- IonQConnection class and notebook, to facilitate experiments through IonQ's API
+- FCISolver active space selection / frozen orbitals: restrictions for half-empty orbitals
+
+### Changed
+
+- QEMIST Cloud MI-FNO innregration: interface adjustments
+- ADAPT-VQE interface: spin only required when needed
+- VQE returns warning or error if no variational parameters are in the ansatz circuit
+- Bug fix: scBK reference state, HEA ansatz initialization
+- Check for valid number of target qubits for common gates
+- Documentation, README
+
+### Deprecated
+
 ## [0.3.1] - 2022-06-15
 
 ### Added